requirements.txt
netifaces
python-dotenv
requests
schedule